pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "protrider"
version='0.9'
description = "PROTRIDER - Protein outlier detection method"
authors = [
    { name = "Daniela Klaproth-Andrade", email = "daniela.andrade@tum.de"},
    { name = "Ines Scheller"},
    { name = "Ata Jadid Ahari"},
    { name = "George Tsitsiridis"},
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "" }
keywords = ["olink", "tools", "ukb", "outliers"]
dependencies = [
    'click',
    'numpy',
    'pandas',
    'pyaml',
    'pathlib',
    'torch',
    'tqdm',
    'optht',
    'pydeseq2',
    'scipy',
    'scikit-learn',
    'pathlib',
    'seaborn',
    'plotnine'
]

[project.urls]
Homepage = "https://github.com/gagneurlab/PROTRIDER"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-xdist",
    "cython"
]

[project.scripts]
protrider = "protrider.cli:cli"
